Refactor quota monitor poll into control plane
