+++ loopx/control_plane/quota/spend_sources.py
@@ -1,0 +1,5 @@
+from __future__ import annotations
+
+
+DEFAULT_SLOT_SPEND_SOURCE = "heartbeat"
+VALID_SLOT_SPEND_SOURCES = {"heartbeat", "controller", "adapter"}
